refactor manager group
