+++ manager.go
@@ -58,12 +58,6 @@ type User struct {
 	IsAdmin  bool   `json:"admin"`
 }
 
-// Group ...
-type Group struct {
-	ID   int    `json:"id"`
-	Name string `json:"name"`
-}
-
 // Session ...
 type Session struct {
 	UserID  int    `json:"id"`
@@ -397,41 +391,4 @@ func (m *Manager) ListUsers(token string) (users []User, err error) {
 		return nil, err
 	}
 	return users, err
-}
-
-// CreateGroup ...
-func (m *Manager) CreateGroup(token string, group string) error {
-	payload := []byte(`{"name": "` + group + `"}`)
-	_, _, err := m.doRequest("/api/groups/", "POST", payload, token, "")
-	if err != nil {
-		fmt.Println(err)
-		return err
-	}
-	color.Green("SUCCESS: Group " + group + " created")
-	return nil
-}
-
-// ListGroups ...
-func (m *Manager) ListGroups(token string) (groups []Group, err error) {
-	body, _, err := m.doRequest("/api/groups/", "GET", []byte(""), token, "")
-	if err != nil {
-		return nil, err
-	}
-	err = json.Unmarshal([]byte(body), &groups)
-	if err != nil {
-		return nil, err
-	}
-	return groups, err
-}
-
-// GroupAddUser ...
-func (m *Manager) GroupAddUser(token string, userid string, groupid string) error {
-	payload := []byte(`{"userid": "` + userid + `", "groupid": "` + groupid + `"}`)
-	_, _, err := m.doRequest("/api/groups/"+groupid+"/users/", "POST", payload, token, "")
-	if err != nil {
-		fmt.Println(err)
-		return err
-	}
-	color.Green("SUCCESS: Added user " + userid + " to group " + groupid)
-	return nil
 }
